# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Habitat Commissioner",
    "description": "Express REST API for moving data between orgs and environments in apigee",
    "license": {
      "name": "MIT",
      "url": "http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user": {
      "get": {
        "description": "Get all information for the logged in user",
        "summary": "Get the logged in user",
        "tags": [
          "user"
        ],
        "operationId": "getUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }
        }
      },
      "put": {
        "description": "Update the user settings for Apigee API hostname, repo parent directory path and which orgs to hit in Apigee",
        "summary": "Update user configuration",
        "tags": [
          "user"
        ],
        "operationId": "updateConfig",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user config",
            "description": "Basic user configuration",
            "schema": {
              "type": "object",
              "properties": {
                "apiHostName": {
                  "type": "string"
                },
                "repoParentDirectory": {
                  "type": "string"
                },
                "orgs": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "in": "header",
            "name": "token",
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }
        }
      },
      "delete": {
        "description": "Delete the logged in user",
        "summary": "Delete the logged in user",
        "tags": [
          "user"
        ],
        "operationId": "deleteUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/repos": {
      "get": {
        "description": "Get the names of each directory in the repo parent directory",
        "summary": "Get repo directories",
        "tags": [
          "user"
        ],
        "operationId": "getRepos",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "description": "",
                  "example": 200,
                  "type": "integer",
                  "format": "int32"
                },
                "message": {
                  "description": "",
                  "example": "Repo directories",
                  "type": "string"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/proxy": {
      "put": {
        "description": "Update proxy settings for the user",
        "summary": "Update proxy settings",
        "tags": [
          "user"
        ],
        "operationId": "updateProxySettings",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "proxy config",
            "description": "Proxy configuration",
            "schema": {
              "type": "object",
              "properties": {
                "enable": {
                  "type": "boolean"
                },
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "scheme": {
                  "type": "string"
                },
                "host": {
                  "type": "string"
                },
                "port": {
                  "type": "string"
                }
              }
            }
          },
          {
            "in": "header",
            "name": "token",
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }
        }
      }
    },
    "/user/ssl": {
      "post": {
        "description": "Enable SSL for the user so they can use the Apigee management API using an SSL key and cert",
        "summary": "Enable SSL",
        "tags": [
          "user"
        ],
        "operationId": "enableSSL",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "passphrase",
            "in": "formData",
            "type": "string",
            "description": "SSL passphrase"
          },
          {
            "name": "key",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "SSL key. Must be a .pem file"
          },
          {
            "name": "cert",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "SSL cert. Must be a .pem file"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access_token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }
        }
      },
      "delete": {
        "description": "Disable SSL for the user. This will also delete any previously uploaded key and cert files from the server",
        "summary": "Disable SSL",
        "tags": [
          "user"
        ],
        "operationId": "disableSSL",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }
        }
      }
    },
    "/user/tests": {
      "post": {
        "description": "Upload a postman collection and environment",
        "summary": "Upload Postman tests",
        "tags": [
          "user"
        ],
        "operationId": "uploadPostmanTest",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "Name for the test suite"
          },
          {
            "name": "collection",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "Postman collection. Must be a .json file"
          },
          {
            "name": "environment",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "Postman environment. Must be a .json file"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }
        }
      }
    },
    "/user/tests/{test_id}": {
      "get": {
        "description": "Get the details of an uploaded postman collection and environment by ID",
        "summary": "Get postman test",
        "tags": [
          "user"
        ],
        "operationId": "getTest",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "test_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Postman test id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "description": "",
                  "example": 200,
                  "type": "integer",
                  "format": "int32"
                },
                "message": {
                  "description": "",
                  "example": "Details for test test_id",
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "collection": {
                      "type": "object",
                      "description": "Postman test collection"
                    },
                    "environment": {
                      "type": "object",
                      "description": "Postman test environment"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Update the collection and/ or environment files for an existing upload",
        "summary": "Update existing postman test",
        "tags": [
          "user"
        ],
        "operationId": "updatePostmanTest",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "Name for the test suite"
          },
          {
            "name": "collection",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "Postman collection. Must be a .json file"
          },
          {
            "name": "environment",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "Postman environment. Must be a .json file"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "test_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The Postman test ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }
        }
      },
      "delete": {
        "description": "Delete a postman collection and environment",
        "summary": "Delete a Postman test",
        "tags": [
          "user"
        ],
        "operationId": "deletePostmanTest",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "test_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The Postman test ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UserResponse": {
      "title": "getUserResponse",
      "example": {
        "code": 200,
        "message": "username",
        "data": {
          "username": "username",
          "apiBasicAuthCredentials": "Basic apigee-access-token",
          "config": {
            "orgs": [
              "org"
            ],
            "repoParentDirectory": "absolute\\path\\to\\repo\\parent",
            "apiHostName": "apigee.management.api.host/v1",
            "ssl": {
              "enable": true,
              "passphrase": "ssl-passphrase",
              "key": "path\\to\\key",
              "cert": "path\\to\\cert"
            },
            "proxy": {
              "enable": true,
              "username": "proxy-username",
              "password": "proxy-password",
              "scheme": "http",
              "host": "proxy.host",
              "port": 8080
            },
            "tests": [
              {
                "id": "test-id",
                "name": "My postman tests",
                "collection": "MyPostmanTest.postman_collection.json",
                "environment": "MyPostmanTest.postman_environment.json"
              }
            ]
          }
        }
      },
      "type": "object",
      "properties": {
        "code": {
          "description": "",
          "example": 200,
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "description": "",
          "example": "username",
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/User"
        }
      },
      "required": [
        "code",
        "message",
        "data"
      ]
    },
...
    "User": {
      "title": "user",
      "example": {
        "username": "username",
        "apiBasicAuthCredentials": "Basic apigee-access-token",
        "config": {
          "orgs": [
            "org"
          ],
          "repoParentDirectory": "absolute\\path\\to\\repo\\parent",
          "apiHostName": "apigee.management.api.host/v1",
          "ssl": {
            "enable": true,
            "passphrase": "ssl-passphrase",
            "key": "path\\to\\key",
            "cert": "path\\to\\cert"
          },
          "proxy": {
            "enable": true,
            "username": "proxy-username",
            "password": "proxy-password",
            "scheme": "http",
            "host": "proxy.host",
            "port": 8080
          },
          "tests": [
            {
              "id": "test-id",
              "name": "My postman tests",
              "collection": "MyPostmanTest.postman_collection.json",
              "environment": "MyPostmanTest.postman_environment.json"
            }
          ]
        }
      },
      "type": "object",
      "properties": {
        "username": {
          "description": "",
          "example": "username",
          "type": "string"
        },
        "apiBasicAuthCredentials": {
          "description": "",
          "example": "Basic apigee-access-token",
          "type": "string"
        },
        "config": {
          "$ref": "#/definitions/UserConfig"
        }
      },
      "required": [
        "username",
        "apiBasicAuthCredentials",
        "config"
      ]
    },
    "UserConfig": {
      "title": "Config",
      "example": {
        "orgs": [
          "org"
        ],
        "repoParentDirectory": "absolute\\path\\to\\repo\\parent",
        "apiHostName": "apigee.management.api.host/v1",
        "ssl": {
          "enable": true,
          "passphrase": "ssl-passphrase",
          "key": "path\\to\\key",
          "cert": "path\\to\\cert"
        },
        "proxy": {
          "enable": true,
          "username": "proxy-username",
          "password": "proxy-password",
          "scheme": "http",
          "host": "proxy.host",
          "port": 8080
        },
        "tests": [
          {
            "id": "test-id",
            "name": "My postman tests",
            "collection": "MyPostmanTest.postman_collection.json",
            "environment": "MyPostmanTest.postman_environment.json"
          }
        ]
      },
      "type": "object",
      "properties": {
        "orgs": {
          "description": "",
          "example": [
            "org"
          ],
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "repoParentDirectory": {
          "description": "",
          "example": "absolute\\path\\to\\repo\\parent",
          "type": "string"
        },
        "apiHostName": {
          "description": "",
          "example": "apigee.management.api.host/v1",
          "type": "string"
        },
        "ssl": {
          "$ref": "#/definitions/UserSsl"
        },
        "proxy": {
          "$ref": "#/definitions/UserProxy"
        },
        "tests": {
          "description": "",
          "example": [
            {
              "id": "test-id",
              "name": "My postman tests",
              "collection": "MyPostmanTest.postman_collection.json",
              "environment": "MyPostmanTest.postman_environment.json"
            }
          ],
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserTest"
          }
        }
      },
      "required": [
        "orgs",
        "repoParentDirectory",
        "apiHostName",
        "ssl",
        "proxy",
        "tests"
      ]
    },
    "UserSsl": {
      "title": "Ssl",
      "example": {
        "enable": true,
        "passphrase": "ssl-passphrase",
        "key": "path\\to\\key",
        "cert": "path\\to\\cert"
      },
      "type": "object",
      "properties": {
        "enable": {
          "description": "",
          "example": true,
          "type": "boolean"
        },
        "passphrase": {
          "description": "",
          "example": "ssl-passphrase",
          "type": "string"
        },
        "key": {
          "description": "",
          "example": "path\\to\\key",
          "type": "string"
        },
        "cert": {
          "description": "",
          "example": "path\\to\\cert",
          "type": "string"
        }
      },
      "required": [
        "enable",
        "passphrase",
        "key",
        "cert"
      ]
    },
    "UserProxy": {
      "title": "Proxy",
      "example": {
        "enable": true,
        "username": "proxy-username",
        "password": "proxy-password",
        "scheme": "http",
        "host": "proxy.host",
        "port": 8080
      },
      "type": "object",
      "properties": {
        "enable": {
          "description": "",
          "example": true,
          "type": "boolean"
        },
        "username": {
          "description": "",
          "example": "proxy-username",
          "type": "string"
        },
        "password": {
          "description": "",
          "example": "proxy-password",
          "type": "string"
        },
        "scheme": {
          "description": "",
          "example": "http",
          "type": "string"
        },
        "host": {
          "description": "",
          "example": "proxy.host",
          "type": "string"
        },
        "port": {
          "description": "",
          "example": 8080,
          "type": "string"
        }
      },
      "required": [
        "enable",
        "username",
        "password",
        "scheme",
        "host",
        "port"
      ]
    },
    "UserTest": {
      "title": "Test",
      "example": {
        "id": "test-id",
        "name": "My postman tests",
        "collection": "MyPostmanTest.postman_collection.json",
        "environment": "MyPostmanTest.postman_environment.json"
      },
      "type": "object",
      "properties": {
        "id": {
          "description": "",
          "example": "test-id",
          "type": "string"
        },
        "name": {
          "description": "",
          "example": "My postman tests",
          "type": "string"
        },
        "collection": {
          "description": "",
          "example": "MyPostmanTest.postman_collection.json",
          "type": "string"
        },
        "environment": {
          "description": "",
          "example": "MyPostmanTest.postman_environment.json",
          "type": "string"
        }
      },
      "required": [
        "id",
        "name",
        "collection",
        "environment"
      ]
    },
...
  }
}